+++ src/frontend/apps/e2e/__tests__/app-impress/404.spec.ts
@@ -3,7 +3,7 @@ import { expect, test } from '@playwright/test';
 test.beforeEach(async ({ page }) => {
   await page.goto('/');
   await expect(
-    page.locator('header').first().locator('h2').getByText('Docs'),
+    page.locator('header').first().locator('h1').getByText('Docs'),
   ).toBeVisible();
   await page.goto('unknown-page404');
 });