+++ src/frontend/apps/e2e/__tests__/app-impress/utils-common.ts
@@ -154,7 +154,7 @@ export const goToGridDoc = async (
   { nthRow = 1, title }: GoToGridDocOptions = {},
 ) => {
   const header = page.locator('header').first();
-  await header.locator('h2').getByText('Docs').click();
+  await header.locator('h1').getByText('Docs').click();
 
   const docsGrid = page.getByTestId('docs-grid');
   await expect(docsGrid).toBeVisible();