+++ src/frontend/apps/impress/src/features/left-panel/components/LeftPanelHeader.tsx
@@ -1,4 +1,5 @@
 import { Button } from '@openfun/cunningham-react';
+import { t } from 'i18next';
 import { useRouter } from 'next/router';
 import { PropsWithChildren, useCallback, useState } from 'react';
 
@@ -54,10 +55,16 @@ export const LeftPanelHeader = ({ children }: PropsWithChildren) => {
             <Box $direction="row" $gap="2px">
               <Button
                 onClick={goToHome}
+                aria-label={t('Back to homepage')}
                 size="medium"
                 color="tertiary-text"
                 icon={
-                  <Icon $variation="800" $theme="primary" iconName="house" />
+                  <Icon
+                    $variation="800"
+                    $theme="primary"
+                    iconName="house"
+                    aria-hidden="true"
+                  />
                 }
               />
               {authenticated && (
@@ -65,8 +72,14 @@ export const LeftPanelHeader = ({ children }: PropsWithChildren) => {
                   onClick={openSearchModal}
                   size="medium"
                   color="tertiary-text"
+                  aria-label={t('Search docs')}
                   icon={
-                    <Icon $variation="800" $theme="primary" iconName="search" />
+                    <Icon
+                      $variation="800"
+                      $theme="primary"
+                      iconName="search"
+                      aria-hidden="true"
+                    />
                   }
                 />
               )}